Fix to AlphabeticalSort

+++ src/AlphabeticalSort.cpp
@@ -5,7 +5,7 @@ class AlphabeticalSort : public BookSort
 	private:
 
 	public:
-		std::vector<Content*> SortBooks(std::unordered_map<int, Content*> catalogue)
+		std::vector<Content*> SortBooks(std::unordered_map<long long, Content*> catalogue)
 		{
 			//This uses the standard library sorting function to sort books alphabetically by title
 			vector<Content*> mapped = MapToVector(catalogue)
